map leader to space bar and move bindings over
Spacebar is a more generic solution for the <leader> key - maybe I'll
move it to , eventually but using the <leader> opens up the possibility
for a lot more customized hotkey workflows and commands for different
types of editing.

The initial project for this is to allow for easy prose / markdown
editing so that I can move my notes + writings into a standard text
editor + workflow.