add ProseMode function and editing customization
This is about making a ProseMode for editing prose and markdown files
easily. Which has different requirements than writing code.

The first part is to add plugins vim-pencil and tagbar. Vim-pencil will
allow movement in softwrapped lines + some convenience editing commands
in insert mode. Tagbar is a ctag extension that opens a pane of ctags
(generated from headers) for easy navigation around the document. These
are fairly specialized plugins so they have a lot of customization.

Next, the ProseMode should be easily toggle-able - so we'll wrap a
function around the settings with some buffer variables to handle state
and make a leader hotkey to switch between modes.